Вынесла форму регистрации в отдельный компонент
